% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pkg.R
\name{assess_pkg}
\alias{assess_pkg}
\title{Assess package}
\usage{
assess_pkg(pkg_source_path, rcmdcheck_args, covr_timeout = Inf)
}
\arguments{
\item{pkg_source_path}{- source path for install local}

\item{rcmdcheck_args}{- arguments for R Cmd check - these come from setup_rcmdcheck_args}

\item{covr_timeout}{- setting for covr time out}
}
\value{
list containing results - list containing metrics, covr, tm - trace matrix, and R CMD check
}
\description{
assess package for risk metrics
}
\keyword{internal}
