% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pkg_r_package.R
\name{assess_pkg_r_package}
\alias{assess_pkg_r_package}
\title{Assess an R package by name (deprecated)}
\usage{
assess_pkg_r_package(package_name, version = NA, repos = getOption("repos"))
}
\arguments{
\item{package_name}{Package name.}

\item{version}{Package version. Default `NA` uses latest.}

\item{repos}{Repository URLs. Default `getOption("repos")`.}
}
\value{
Assessment results or `NULL`.
}
\description{
Use [risk_assess_pkg()] with `package` instead.
}
