% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_risk_profile}
\alias{calc_risk_profile}
\title{Calculate package risk profile}
\usage{
calc_risk_profile(risk_data)
}
\arguments{
\item{risk_data}{overall risk score}
}
\value{
risk_profile
}
\description{
Calculate package risk profile
}
