% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_exists_and_overwrite}
\alias{check_exists_and_overwrite}
\title{Check if a path exists and delete the file
if overwrite is TRUE}
\usage{
check_exists_and_overwrite(path, overwrite)
}
\arguments{
\item{path}{a file path to check if it exists}

\item{overwrite}{Logical (T/F). If `TRUE`, delete the file at the specified path}
}
\description{
Check if a path exists and delete the file
if overwrite is TRUE
}
