% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_riskscore_data}
\alias{check_riskscore_data}
\title{check if risk score data exists}
\usage{
check_riskscore_data(current_package)
}
\arguments{
\item{current_package}{- current risk assessment package}
}
\value{
riskscore_data_list - list with path and exists logical
}
\description{
check if risk score data exists
}
