% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_riskscore_data_location}
\alias{check_riskscore_data_location}
\title{check if risk score data exists}
\usage{
check_riskscore_data_location(results_dir)
}
\arguments{
\item{results_dir}{- path to the results directory}
}
\value{
riskscore_data_list - list with path and exists logical
}
\description{
check if risk score data exists
}
