% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports.R
\name{classify_function_body}
\alias{classify_function_body}
\title{classify_function_body}
\usage{
classify_function_body(row, package_name)
}
\arguments{
\item{row}{- row with exported functions}

\item{package_name}{- name of the package}
}
\value{
row - row with the function details plus function body
}
\description{
This function gets the function bodies for exported functions
}
\keyword{internal}
