% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{clean_up_dependencies}
\alias{clean_up_dependencies}
\title{Clean Up Dependencies}
\usage{
clean_up_dependencies(x)
}
\arguments{
\item{x}{A character vector of dependency strings.}
}
\value{
A character vector of unique package names.
}
\description{
Cleans up a list of dependency strings by extracting unique package names.
}
\keyword{internal}
