% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_vignette_folder.R
\name{contains_vignette_folder}
\alias{contains_vignette_folder}
\title{Check for Vignette Folder and .Rmd Files in a .tar File}
\usage{
contains_vignette_folder(tar_file)
}
\arguments{
\item{tar_file}{A character string specifying the path to the .tar file to be checked.}
}
\value{
A logical value: \code{TRUE} if the 'vignette' folder exists and contains at least one .Rmd file,
\code{FALSE} otherwise.
}
\description{
This function checks if a given .tar file contains a 'vignette' folder
and if there is at least one .Rmd file within that folder.
}
\examples{
\dontrun{
  tar_file <- "path/to/your/package.tar.gz"
  result <- contains_vignette_folder(tar_file)
  print(result)
}

}
