% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download.R
\name{convert_abbreviation_to_number}
\alias{convert_abbreviation_to_number}
\title{Convert Abbreviated Numbers to Integers}
\usage{
convert_abbreviation_to_number(value)
}
\arguments{
\item{value}{A character string representing a number (e.g., '1.2M', '500K', '100').}
}
\value{
An integer representation of the numeric value.
}
\description{
Converts a string containing a number with optional 'K' or 'M' abbreviation into an integer.
}
\examples{
\dontrun{
convert_abbreviation_to_number('1.2M') 
# 1200000
convert_abbreviation_to_number('500K') 

convert_abbreviation_to_number('100')
# 100
}
}
\keyword{internal}
