% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{convert_number_to_abbreviation}
\alias{convert_number_to_abbreviation}
\title{Convert number to abbreviation}
\usage{
convert_number_to_abbreviation(value)
}
\arguments{
\item{value}{- number to be converted}
}
\description{
Convert number to abbreviation
}
\keyword{internal}
