% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{expand_dependency_type_spec}
\alias{expand_dependency_type_spec}
\title{Expand Dependency Type Specification}
\usage{
expand_dependency_type_spec(x)
}
\arguments{
\item{x}{A character string specifying the dependency type ("strong", "most", "all", or custom).}
}
\value{
A character vector of dependency types.
}
\description{
Expands a dependency type specification into a character vector of dependency types.
}
\keyword{internal}
