% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{extract_coverage_df}
\alias{extract_coverage_df}
\title{Build a file-coverage data frame from a single res_cov object}
\usage{
extract_coverage_df(res_cov, pkg_name)
}
\arguments{
\item{res_cov}{The \code{res_cov} list (containing \code{coverage}, \code{errors}, \code{notes}).}

\item{pkg_name}{Name of the package (used to strip temp paths from file names).}
}
\description{
Build a file-coverage data frame from a single res_cov object
}
\keyword{internal}
