% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tree.R
\name{extract_license_from_description}
\alias{extract_license_from_description}
\title{Extract License from Package DESCRIPTION File}
\usage{
extract_license_from_description(path)
}
\arguments{
\item{path}{A character string specifying the path to the package directory containing the DESCRIPTION file.}
}
\value{
A character string representing the package license, or `NA` if not found.
}
\description{
This function extracts the license information from a package's DESCRIPTION file.
}
\keyword{internal}
