% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{extract_maintainer_info}
\alias{extract_maintainer_info}
\title{Helper to create maintainer}
\usage{
extract_maintainer_info(person_obj)
}
\arguments{
\item{person_obj}{- value}
}
\description{
Helper to create maintainer
}
\keyword{internal}
