% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tree.R
\name{extract_package_version}
\alias{extract_package_version}
\title{Extract the Installed Version of a Package}
\usage{
extract_package_version(package_name)
}
\arguments{
\item{package_name}{A character string specifying the name of the package.}
}
\value{
A character string representing the installed package version, or `NA` if the package is not found.
}
\description{
This function retrieves the installed version of a specified R package.
If the package is not installed, returns `NA`.
}
\examples{
\dontrun{
extract_package_version("dplyr")
# 1.1.4
}

}
