% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_analysis.R
\name{extract_risk_inputs}
\alias{extract_risk_inputs}
\title{Extract Risk Inputs}
\usage{
extract_risk_inputs(flat_data)
}
\arguments{
\item{flat_data}{Flat package data as returned by \code{normalize_data()}.}
}
\value{
A list of computed values required for risk scoring.
}
\description{
Compute intermediate risk input values from flattened package data.
}
\keyword{internal}
