% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thresholds.R
\name{extract_thresholds_by_id}
\alias{extract_thresholds_by_id}
\title{Extract risk thresholds by id}
\usage{
extract_thresholds_by_id(risk_rule, id_value)
}
\arguments{
\item{risk_rule}{- list containing all risk rules}

\item{id_value}{- character vector containing risk rule choice}
}
\value{
list containing specific risk rules
}
\description{
Extract risk thresholds by id
}
