% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thresholds.R
\name{extract_thresholds_by_key}
\alias{extract_thresholds_by_key}
\title{Extract risk thresholds by key}
\usage{
extract_thresholds_by_key(risk_rule, key_value)
}
\arguments{
\item{risk_rule}{- list containing all risk rules}

\item{key_value}{- character vector containing risk rule choice}
}
\value{
list containing specific risk rules
}
\description{
Extract risk thresholds by key
}
