% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioconductor_data.R
\name{fetch_bioconductor_package_info}
\alias{fetch_bioconductor_package_info}
\title{Fetch Bioconductor Package Information}
\usage{
fetch_bioconductor_package_info(bioconductor_version, package_name)
}
\arguments{
\item{bioconductor_version}{A character string specifying the Bioconductor version (e.g., "3.14").}

\item{package_name}{A character string specifying the name of the package.}
}
\value{
A list containing package details, including the latest version, package URL,
        source package link, and any archived versions if available.
        Returns `FALSE` if the package does not exist or cannot be retrieved.
}
\description{
This function retrieves information about a specific Bioconductor package
for a given Bioconductor version. It fetches the package details, such as
version, source package URL, and archived versions if available.
}
\examples{
\dontrun{
fetch_bioconductor_package_info("3.14", "GenomicRanges")

[[1]]
[[1]]$version
[1] "1.46.1"

[[1]]$url
[1] "https://bioconductor.org/packages/GenomicRanges"

[[1]]$source_package
[1] "https://bioconductor.org/packages/3.14/bioc/src/contrib/GenomicRanges_1.46.1.tar.gz"

[[1]]$bioconductor_version
[1] "3.14"

[[1]]$url_package
[1] "https://bioconductor.org/packages/GenomicRanges"

[[1]]$archived
[1] FALSE


[[2]]
[[2]]$version
NULL

[[2]]$url
[1] "https://bioconductor.org/packages/GenomicRanges"

[[2]]$source_package
[1] "https://bioconductor.org/packages/3.14/bioc/src/contrib/Archive/GenomicRanges/"

[[2]]$archived
[1] TRUE

[[2]]$bioconductor_version
[1] "3.14"
}
}
