% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R
\name{fine_grained_tms}
\alias{fine_grained_tms}
\title{Create fine grained traceability matrices}
\usage{
fine_grained_tms(tm, pkg_name)
}
\arguments{
\item{tm}{- general traceability matrix with no filtered information}

\item{pkg_name}{- name of the package}
}
\value{
- nested list with fine grained tms
}
\description{
Create fine grained traceability matrices
}
\keyword{internal}
