% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_fg_trace_matrix_section}
\alias{generate_fg_trace_matrix_section}
\title{Generate Fine grained Trace Matrices Section}
\usage{
generate_fg_trace_matrix_section(tm_type, tm_df)
}
\arguments{
\item{tm_type}{- type of traceability matrix}

\item{tm_df}{- input data}
}
\description{
Generates the Trace Matrices section for the HTML report.
}
\keyword{internal}
