% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_pop_metrics_section}
\alias{generate_pop_metrics_section}
\title{Generate Popularity Metrics Section}
\usage{
generate_pop_metrics_section(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the popularity section for the HTML report.
}
\keyword{internal}
