% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_report.R
\name{generate_risk_analysis}
\alias{generate_risk_analysis}
\title{Generate Risk Analysis}
\usage{
generate_risk_analysis(assessment_results)
}
\arguments{
\item{assessment_results}{- input data containing risk analysis metrics}
}
\description{
Generates the Risk Analysis table for the HTML report.
}
\keyword{internal}
