% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports.R
\name{get_all_s4_methods}
\alias{get_all_s4_methods}
\title{Extract all S4 methods}
\usage{
get_all_s4_methods(generic_name)
}
\arguments{
\item{generic_name}{name of function}
}
\value{
list with method_strings, generic, and methods
}
\description{
Extract all S4 methods
}
\keyword{internal}
