% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get dependencies}
\usage{
get_dependencies(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{package source path}
}
\description{
Get dependencies
}
\keyword{internal}
