% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R
\name{get_exports}
\alias{get_exports}
\title{list all package exports}
\usage{
get_exports(pkg_source_path)
}
\value{
data.frame, with one column `exported_function`, that can be passed
  to all downstream map_* helpers
}
\description{
list all package exports
}
\keyword{internal}
