% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{get_max_thresholds}
\alias{get_max_thresholds}
\title{Extract Maximum Thresholds for Code Coverage Levels}
\usage{
get_max_thresholds(thresholds)
}
\arguments{
\item{thresholds}{A list of threshold objects, each containing a `level` and `max` field.}
}
\value{
A list with three elements:
\describe{
  \item{high_max}{The maximum threshold value for the `"high"` level.}
  \item{medium_max}{The maximum threshold value for the `"medium"` level.}
  \item{medium_max}{The maximum threshold value for the `"low"` level.}
}
}
\description{
This internal function retrieves the `max` values for `"high"`,`"medium"`, and `"low"` levels
from a list of code coverage thresholds.
}
\keyword{internal}
