% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_version.R
\name{get_package_version}
\alias{get_package_version}
\title{Get Package Version Information}
\usage{
get_package_version(package_name, last_num = 10)
}
\arguments{
\item{package_name}{The name of the R package for which version information is requested.}

\item{last_num}{The number of most recent versions to return, default is 10.}
}
\value{
If successful, returns a tibble with columns `package_name`, `package_version`,
`link`, `date`, and `size` indicating the version details of the package. The results are sorted
by the most recent dates. If the specified package name does not exist in the response or if the
API call fails, the function may return NULL or an error message, depending on the API's response.
}
\description{
Retrieves the version information of a specified R package by querying a REST API.
This function sends an HTTP POST request to the specified URL and expects a JSON response
containing the version details of the requested package. It returns the most recent versions
based on the 'last_num' parameter.
}
\examples{
\dontrun{
# Example usage:
# Assuming the function and the API endpoint are properly set up and accessible:
result <- sanofi.risk.metric::get_package_version("dplyr")
print(result)
# Expected result (truncated):
# A tibble: 44 × 5
#   package_name    package_version link                 date             size 
#  <chr>           <chr>           <chr>                <chr>            <chr>
# 1 dplyr           0.1.1           dplyr_0.1.1.tar.gz   2014-01-29 21:24 530K 
# 2 dplyr           0.1.2           dplyr_0.1.2.tar.gz   2014-02-24 16:36 533K 
# 3 dplyr           0.1.3           dplyr_0.1.3.tar.gz   2014-03-15 00:36 535K 
# 4 dplyr           0.1             dplyr_0.1.tar.gz     2014-01-16 22:53 2.7M 
}
}
