% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports.R
\name{get_r6_methods_details}
\alias{get_r6_methods_details}
\title{function to get R6 methods}
\usage{
get_r6_methods_details(class_name)
}
\arguments{
\item{class_name}{- function class}
}
\description{
function to get R6 methods from class
}
\keyword{internal}
