% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_reverse_dependencies}
\alias{get_reverse_dependencies}
\title{Get reverse dependencies}
\usage{
get_reverse_dependencies(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{package source path}
}
\description{
Get reverse dependencies
}
\keyword{internal}
