% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_skip_stf.R
\name{get_tests_skip_stf}
\alias{get_tests_skip_stf}
\title{get skipped tests with line locations}
\usage{
get_tests_skip_stf(problem_tests, test_map)
}
\arguments{
\item{problem_tests}{- problem tests}

\item{test_map}{- all tests with line locations}
}
\value{
test_skip
}
\description{
get skipped tests with line locations
}
\keyword{internal}
