% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_local.R
\name{install_package_local}
\alias{install_package_local}
\title{install package locally with source}
\usage{
install_package_local(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{directory path to R project}
}
\description{
install package locally with source
}
