% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tree.R
\name{is_base}
\alias{is_base}
\title{Check if a Package is a Base or Recommended R Package}
\usage{
is_base(pkg)
}
\arguments{
\item{pkg}{A character string representing the name of the package.}
}
\value{
A logical value: `TRUE` if the package is a base or recommended package, `FALSE` otherwise.
}
\description{
This function determines whether a given package is a base or recommended package
using CRAN metadata.
}
\examples{
\dontrun{
is_base("stats")     # TRUE
is_base("ggplot2")   # FALSE
}
}
