% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_description_file.R
\name{modify_description_file}
\alias{modify_description_file}
\title{Modify the DESCRIPTION File in a R Package Tarball}
\usage{
modify_description_file(tar_file, package_name)
}
\arguments{
\item{tar_file}{A string representing the path to the `.tar.gz` file that contains the R package.}

\item{package_name}{A string representing the name of the package, which should match the directory name after extraction.}
}
\value{
A string containing the path to the newly created modified `.tar.gz` file.
}
\description{
This function recreate a `.tar.gz` R package file after modifying its `DESCRIPTION` file
by appending Config/build/clean-inst-doc: false parameter.
}
\examples{
\dontrun{
  modified_tar <- modify_description_file("path/to/mypackage.tar.gz", "mypackage")
  print(modified_tar)
}

}
