% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{rd_aliases}
\alias{rd_aliases}
\title{Extract all aliases from an Rd document}
\usage{
rd_aliases(rd)
}
\arguments{
\item{rd}{An Rd object.}
}
\value{
A character vector of aliases (possibly empty).
}
\description{
Collects text from **all** `\\alias{}` nodes, trimming whitespace and
de-duplicating.
}
\keyword{internal}
