% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{rd_flatten_text}
\alias{rd_flatten_text}
\title{Flatten an Rd node to plain text}
\usage{
rd_flatten_text(node)
}
\arguments{
\item{node}{An Rd node or list of nodes.}
}
\value{
Character scalar containing the concatenated text.
}
\description{
Recursively flattens a node (or list of nodes) into a character string.
}
\keyword{internal}
