% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{rd_name}
\alias{rd_name}
\title{Extract the topic name from an Rd document}
\usage{
rd_name(rd)
}
\arguments{
\item{rd}{An Rd object.}
}
\value{
A length-1 character scalar (topic name) or `NA_character_` if missing.
}
\description{
Returns the `\\name{}` (roxygen `@rdname`) of the Rd document.
}
\keyword{internal}
