% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recalc_risk_scores}
\alias{recalc_risk_scores}
\title{Re-calculate package risk scores}
\usage{
recalc_risk_scores(riskdata_results, update_comments)
}
\arguments{
\item{riskdata_results}{- path to riskdata_results toy dataset}

\item{update_comments}{notes explaining why score recalculated}
}
\description{
{Use this function to re-calculate risk scores and risk profile}
}
\details{
{Use cases: if the weighting profile and/or risk profile thresholds
have changed and the risk metrics have not changed, then
use this function to re-calculate the risk scores and profile
without running the whole risk assessment process again}
}
\examples{
riskdata_results <- system.file("test-data/riskdata_results_slim.csv", 
package = "risk.assessr")
update_comments <- "recalc scores example"

recalc_example <- recalc_risk_scores(riskdata_results, update_comments)

}
