% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covr_skip_nstf.R
\name{run_covr_skip_nstf}
\alias{run_covr_skip_nstf}
\title{Run Coverage Analysis on Package Source and Tests}
\usage{
run_covr_skip_nstf(pkg_source_path, test_pkg_data)
}
\arguments{
\item{pkg_source_path}{Character string. Path to the root of the package source directory.}

\item{test_pkg_data}{List. Output from `check_pkg_tests_and_snaps()` indicating test framework presence.}
}
\value{
returns a list with coverage object and source files with no tests.
}
\description{
This internal function maps source files to test files, checks for skipped/problematic tests,
runs the remaining test files manually, calculates coverage, and optionally generates a report.
}
\examples{
\dontrun{
r = getOption("repos")
r["CRAN"] = "http://cran.us.r-project.org"
options(repos = r)

dp <- system.file("test-data", 
  "MASS_7.3-65.tar.gz", 
  package = "risk.assessr")

# set up package
install_list <- set_up_pkg(dp)

build_vignettes <- install_list$build_vignettes
package_installed <- install_list$package_installed
pkg_source_path <- install_list$pkg_source_path
rcmdcheck_args <- install_list$rcmdcheck_args

# install package locally to ensure test works
package_installed <- install_package_local(pkg_source_path)
package_installed <- TRUE

rcmdcheck_args$path <- pkg_source_path
test_pkg_data <- check_pkg_tests_and_snaps(pkg_source_path)
test_covr_list <- run_covr_skip_nstf(pkg_source_path, test_pkg_data)

options(old)
}
}
