% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanofi.risk.metric-package.R
\docType{package}
\name{sanofi.risk.metric-package}
\alias{sanofi.risk.metric}
\alias{sanofi.risk.metric-package}
\title{sanofi.risk.metric: Sanofi risk metric}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sanofi riskmetric package.
}
\author{
\strong{Maintainer}: Edward Gillian \email{edward.gillian-ext@sanofi.com}

Authors:
\itemize{
  \item Hugo Bottois \email{hugo.bottois@sanofi.com}
}

}
\keyword{internal}
