% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_pkg.R
\name{set_up_pkg}
\alias{set_up_pkg}
\title{Creates information on package installation}
\usage{
set_up_pkg(package_path, check_type = "1")
}
\arguments{
\item{package_path}{Path to the package tarball (.tar.gz).}

\item{check_type}{Type of R CMD check: "1" for basic, "2" for CRAN-like.}
}
\value{
list with local package install
}
\description{
Creates information on package installation
}
\examples{
\dontrun{
set_up_pkg("path/to/tarball")
}
}
