% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R
\name{width_adjust_func}
\alias{width_adjust_func}
\title{adjust width of columns}
\usage{
width_adjust_func(wb, DF, sheetName)
}
\arguments{
\item{wb}{- openxlsx workbook}

\item{DF}{- dataframe}

\item{sheetName}{- name of Excel sheet}
}
\description{
adjust width of columns to either values in df or column headers
}
\keyword{internal}
