% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_data_csv}
\alias{write_data_csv}
\title{Write results to csv}
\usage{
write_data_csv(data, riskscore_data_path, riskscore_data_exists)
}
\arguments{
\item{data}{- results data}

\item{riskscore_data_path}{directory path and file name}

\item{riskscore_data_exists}{logical with T/F if risk score data exists}
}
\description{
Write results to csv
}
