% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pkg.R
\name{write_pkg}
\alias{write_pkg}
\title{write package}
\usage{
write_pkg(assess_package, results_dir, overwrite = TRUE)
}
\arguments{
\item{assess_package}{- nested list with risk metrics}

\item{results_dir}{- path for writing results}

\item{overwrite}{Logical (T/F). Whether or not to overwrite existing risk metric results}
}
\description{
write package data for risk metrics
}
\examples{
\dontrun{
library(sanofi.risk.metric)
# set CRAN repo 
r = getOption("repos")
r["CRAN"] = "http://cran.us.r-project.org"
options(repos = r)

dp <- system.file("test-data/test.package.0001_0.1.0.tar.gz",
                  package = "sanofi.risk.metric")

# set up package
install_list <- sanofi.risk.metric::set_up_pkg(dp)

build_vignettes <- install_list$build_vignettes
package_installed <- install_list$package_installed
pkg_source_path <- install_list$pkg_source_path
rcmdcheck_args <- install_list$rcmdcheck_args

if (package_installed == TRUE ) {
   
  assess_package <- 
    sanofi.risk.metric::assess_pkg(pkg_source_path,
                                   rcmdcheck_args)
  
  assess_package$results$comments <- "test run"
                                    
  sanofi.risk.metric::write_pkg(assess_package, 
                                results_dir = "no audit trail")                                  
}
}

}
