% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R
\name{write_tm_rds}
\alias{write_tm_rds}
\title{Write tm to rds}
\usage{
write_tm_rds(tm, pkg_name, results_dir)
}
\arguments{
\item{tm}{- traceability matrix}

\item{pkg_name}{- name of the package}

\item{results_dir}{- directory where rds is written}
}
\description{
This function writes the traceability matrix to rds
}
\keyword{internal}
