% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pkg.R
\name{assess_pkg}
\alias{assess_pkg}
\title{Assess package}
\usage{
assess_pkg(pkg_source_path, rcmdcheck_args, covr_timeout = Inf)
}
\arguments{
\item{pkg_source_path}{- source path for install local}

\item{rcmdcheck_args}{- arguments for R Cmd check - these come from setup_rcmdcheck_args}

\item{covr_timeout}{- setting for covr time out}
}
\value{
list containing results - list containing metrics, covr, tm - trace matrix, and R CMD check
}
\description{
assess package for risk metrics
}
\examples{
\dontrun{
library(sanofi.risk.metric)
# set CRAN repo 
r = getOption("repos")
r["CRAN"] = "http://cran.us.r-project.org"
options(repos = r)

dp <- system.file("test-data/test.package.0001_0.1.0.tar.gz",
                  package = "sanofi.risk.metric")

# set up package
install_list <- sanofi.risk.metric::set_up_pkg(dp)

build_vignettes <- install_list$build_vignettes
package_installed <- install_list$package_installed
pkg_source_path <- install_list$pkg_source_path
rcmdcheck_args <- install_list$rcmdcheck_args

if (package_installed == TRUE ) {
   
  assess_package <- 
    sanofi.risk.metric::assess_pkg(pkg_source_path,
                                   rcmdcheck_args)
}
}  
}
