% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pkg.R
\name{assess_pkg}
\alias{assess_pkg}
\title{Assess package}
\usage{
assess_pkg(
  pkg,
  datapath,
  pkg_source_path,
  out_dir,
  riskscore_data_path,
  riskscore_data_exists,
  overwrite = TRUE,
  rcmdcheck_args = list(timeout = Inf, args = c("--ignore-vignettes", "--no-vignettes",
    "--no-manual"), build_args = "--no-build-vignettes", env =
    c(`_R_CHECK_FORCE_SUGGESTS_` = "FALSE"), quiet = TRUE),
  covr_timeout = Inf
)
}
\arguments{
\item{pkg}{- name of the package}

\item{datapath}{- datapath with 0.gz file}

\item{pkg_source_path}{- source path for install local}

\item{out_dir}{path for writing results}

\item{riskscore_data_path}{data path of current risk assessment package}

\item{riskscore_data_exists}{logical with T/F if risk score data exists}

\item{overwrite}{Logical (T/F). Whether or not to overwrite existing scorecard results}

\item{rcmdcheck_args}{- arguments for R Cmd check}

\item{covr_timeout}{- setting for covr time out}
}
\value{
results - list containing metrics
}
\description{
assess package for risk metrics
}
