% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pkg.R
\name{assess_pkg}
\alias{assess_pkg}
\title{Assess package}
\usage{
assess_pkg(
  pkg,
  datapath,
  pkg_source_path,
  out_dir,
  overwrite = TRUE,
  rcmdcheck_args = list(timeout = Inf, args = "--no-manual", quiet = TRUE),
  covr_timeout = Inf
)
}
\arguments{
\item{pkg}{- name of the package}

\item{datapath}{- datapath with 0.gz file}

\item{pkg_source_path}{- source path for install local}

\item{out_dir}{path for writing results}

\item{overwrite}{Logical (T/F). Whether or not to overwrite existing scorecard results}

\item{rcmdcheck_args}{- arguments for R Cmd check}

\item{covr_timeout}{- setting for covr time out}
}
\value{
results - list containing metrics
}
\description{
assess package for risk metrics
}
