% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_overall_risk_score}
\alias{calc_overall_risk_score}
\title{Calculate overall package risk scores}
\usage{
calc_overall_risk_score(data, weights = NULL)
}
\arguments{
\item{data}{risk metric data}

\item{weights}{vector with weight values}
}
\value{
pkg_score
}
\description{
Calculate overall package risk scores
}
