% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tree.R
\name{fetch_all_dependencies}
\alias{fetch_all_dependencies}
\title{Fetch All Dependencies for a Package}
\usage{
fetch_all_dependencies(package_name)
}
\arguments{
\item{package_name}{A character string specifying the name of the package.}
}
\value{
A nested list representing the dependency tree of the package.
}
\description{
This function builds and retrieves the full dependency tree for a given R package.
}
\examples{
\dontrun{
fetch_all_dependencies("ggplot2")
}

}
