% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_Bioconductor_Package_Info.R
\name{fetch_Bioconductor_Package_Info}
\alias{fetch_Bioconductor_Package_Info}
\title{Get Bioconductor Package Version Information}
\usage{
fetch_Bioconductor_Package_Info(package_name, version)
}
\arguments{
\item{package_name}{The name of the Bioconductor package for which version information is requested.}

\item{version}{The specific version of the package you want information about.}
}
\value{
If successful, returns a list containing version details of the package. The structure
of the response depends on the API's response format. If the specified package name does not exist
in the response or if the API call fails, the function may return NULL or an error message,
depending on the API's response.
}
\description{
Retrieves the version information of a specified Bioconductor package by querying a REST API.
This function sends an HTTP POST request to the specified URL and expects a JSON response
containing the version details of the requested package.
}
\examples{
\dontrun{
# Example usage:
package_info <- sanofi.risk.assessr::fetch_Bioconductor_Package_Info("GenomicRanges", "1.42.0")
print(package_info)
}
}
