% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tar_package.R
\name{get_tar_package}
\alias{get_tar_package}
\title{Get TAR Package URL}
\usage{
get_tar_package(package_name, version = NA)
}
\arguments{
\item{package_name}{A character string specifying the name of the package.}

\item{version}{A character string specifying the version of the package.}
}
\value{
A list containing the response from the API, which includes the URL of the TAR package.
}
\description{
This function retrieves the URL of a TAR package given its name and version.
}
\details{
This function sends a POST request to a specified API endpoint with the package name and version as the body. 
If the request is successful, the function returns the response from the API, which contains the URL of the TAR package.
If the request fails, the function stops and returns an error message with the status code.
}
\examples{
\dontrun{
result <- get_tar_package("dplyr", "1.0.0")
print(result)
# Expected output:
$package_name
[1] "dplyr"

$version
[1] "1.0.0"

$tar_link
[1] "https://cran.r-project.org/src/contrib/Archive/dplyr/dplyr_1.0.0.tar.gz"

$source
[1] "CRAN"

$error
NULL

$version_available
[1] "0.1.1, 0.1.2, 0.1.3, 0.1, 0.2, 0.3.0.1, 0.3.0.2, 0.3, 0.4.0, 0.4.1, 
0.4.2, 0.4.3, 0.5.0, 0.7.0, 0.7.1, 0.7.2, 0.7.3, 0.7.4, 0.7.5, 0.7.6, 
0.7.7, 0.7.8, 0.8.0.1, 0.8.0, 0.8.1, 0.8.2, 0.8.3, 0.8.4, 0.8.5, 1.0.0, 
1.0.1, 1.0.2, 1.0.3, 1.0.4, 1.0.5, 1.0.6, 1.0.7, 1.0.8, 1.0.9, 1.0.10, 
1.1.0, 1.1.1, 1.1.2, 1.1.3, 1.1.4"
}
}
