% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_local.R
\name{install_package_local}
\alias{install_package_local}
\title{install package locally with source}
\usage{
install_package_local(pkg_source_path, pkg_disp)
}
\arguments{
\item{pkg_source_path}{directory path to R project}

\item{pkg_disp}{vector with package name for display}
}
\description{
install package locally with source
}
