% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.assessr-package.R
\docType{package}
\name{risk.assessr-package}
\alias{risk.assessr}
\alias{risk.assessr-package}
\title{risk.assessr: Assessing Package Risk Metrics}
\description{
A reliable and validated tool that captures detailed risk metrics such as R 'CMD' check, test coverage, traceability matrix, documentation, dependencies, reverse dependencies, suggested dependency analysis, repository data, and enhanced reporting for R packages that are local or stored on remote repositories such as GitHub, CRAN, and Bioconductor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/risk.assessr/}
  \item Report bugs at \url{https://github.com/pharmaverse/risk.assessr/issues}
}

}
\author{
\strong{Maintainer}: Edward Gillian \email{edward.gillian-ext@sanofi.com} (\href{https://orcid.org/0000-0003-2732-5107}{ORCID})

Authors:
\itemize{
  \item Hugo Bottois \email{hugo.bottois@sanofi.com} (\href{https://orcid.org/0000-0003-4674-0875}{ORCID})
  \item Paulin Charliquart \email{paulin.charliquart@sanofi.com}
  \item Andre Couturier \email{andre.couturier@sanofi.com}
}

Other contributors:
\itemize{
  \item Sanofi [copyright holder, funder]
}

}
\keyword{internal}
