% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.assessr-package.R
\docType{package}
\name{risk.assessr-package}
\alias{risk.assessr}
\alias{risk.assessr-package}
\title{risk.assessr: Assessing package risk metrics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for assessing R packages against a number of metrics to help validate their trustworthiness and reliability.
}
\author{
\strong{Maintainer}: Edward Gillian \email{edward.gillian-ext@sanofi.com} (\href{https://orcid.org/0000-0003-2732-5107}{ORCID})

Authors:
\itemize{
  \item Hugo Bottois \email{hugo.bottois@sanofi.com} (\href{https://orcid.org/0000-0003-4674-0875}{ORCID})
  \item Paulin Charliquart \email{paulin.charliquart@sanofi.com}
  \item Andre Couturier \email{andre.couturier@sanofi.com}
}

Other contributors:
\itemize{
  \item Sanofi [copyright holder, funder]
}

}
\keyword{internal}
