% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.metric-package.R
\docType{package}
\name{risk.assessr-package}
\alias{risk.assessr}
\alias{risk.assessr-package}
\title{risk.assessr: Assessing package risk metrics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for assessing R packages against a number of metrics to help validate their trustworthiness and reliability.
}
\author{
\strong{Maintainer}: Edward Gillian \email{edward.gillian-ext@sanofi.com}

Authors:
\itemize{
  \item Hugo Bottois \email{hugo.bottois@sanofi.com}
}

}
\keyword{internal}
