% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_pkg.R
\name{set_up_pkg}
\alias{set_up_pkg}
\title{Creates information on package installation}
\usage{
set_up_pkg(dp, pkg_disp)
}
\arguments{
\item{dp}{data path and name for the package.}

\item{pkg_disp}{package name for display}
}
\value{
list with local package install
}
\description{
Creates information on package installation
}
